requirements.txt
Django==5.2.12
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==12.1.1
openai==2.28.0
